Rename debug to data faker
